init supplier domain

+++ internal/domain/supplier/entity/relation.go
@@ -6,7 +6,7 @@ import (
 
 type Relation struct {
 	xdb.Entity
-	ContactId int    `gorm:"column:contact_id;NOT NULL" json:"contact_id"`
+	ContactId int64  `gorm:"column:contact_id;NOT NULL" json:"contact_id"`
 	Type      int    `gorm:"column:type;NOT NULL" json:"type"`   // 联系方式类型
 	Value     string `gorm:"column:value;NOT NULL" json:"value"` // 联系方式值
 	xdb.CUDTime
